requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0